Splits walk functions out from the rest of the template code.
